+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 - Logowanie, rejestracja, przypomnienie hasła
 - Szablon strony
 - Menu
-- Lista zwierząt
-- Lista schronisk
+- Lista zwierząt (/pets)
+- Lista schronisk (/shelters)
+- Administracyjna lista zwierząt (/admin/pets) + przyciski do narzędzi
+- Administracyjna lista typów zwierząt (/admin/pet-types)
+- Administracyjna lista typów aktywności (/admin/action-types)
+- Administracyjna lista użytkowników (/admin/users)
+- Administracyjna lista artykułów (/admin/articles)
+- Administracyjna lista komentarzy (/admin/comments/:articleId)
+- Menu administracyjne